% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement.R
\docType{data}
\name{K_DISPLACEMENT}
\alias{K_DISPLACEMENT}
\title{Scaling constant for displacement (in degrees)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
K_DISPLACEMENT
}
\description{
Scaling constant for displacement (in degrees)
}
\keyword{datasets}
