% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_utils.R
\name{ssm_effect_label}
\alias{ssm_effect_label}
\title{Effect Size Label}
\usage{
ssm_effect_label(effect, type = c("amplitude", "elevation"))
}
\arguments{
\item{effect}{Effect size value}

\item{type}{Either "amplitude" or "elevation"}
}
\value{
Character label describing effect size magnitude
}
\description{
Interprets effect size based on Zimmermann and Wright (2017) benchmarks
}
\details{
Benchmarks from Zimmermann & Wright (2017) based on 433 SSM analyses:
\itemize{
\item Amplitude: Small = 0.10, Medium = 0.16, Large = 0.23
\item Elevation: Small = 0.02, Medium = 0.11, Large = 0.27
}
}
\examples{
ssm_effect_label(0.12, "amplitude")
ssm_effect_label(0.25, "elevation")

}
