% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_parameters.R
\name{ssm_parameters}
\alias{ssm_parameters}
\title{Calculate SSM Parameters from Octant Scores}
\usage{
ssm_parameters(scores, angles = seq(0, 315, by = 45))
}
\arguments{
\item{scores}{Numeric vector of length 8 (octant scores/correlations)}

\item{angles}{Numeric vector of octant angles in degrees (default: 0, 45, ..., 315)}
}
\value{
An object of class \code{ssm_params} containing:
\describe{
\item{elevation}{Mean of the octant scores}
\item{amplitude}{Distance from origin to profile center}
\item{displacement_deg}{Angular location in degrees (0-360)}
\item{displacement_rad}{Angular location in radians}
\item{x_val}{X coordinate (Dominance axis)}
\item{y_val}{Y coordinate (Love/Affiliation axis)}
\item{scores}{Original input scores}
\item{angles}{Original input angles}
}
}
\description{
Computes elevation, amplitude, and displacement from a vector of 8 octant
scores (typically correlations with an external variable).
}
\examples{
# Example: correlations of external variable with 8 IPC octants
octant_cors <- c(0.4, 0.3, 0.1, -0.1, -0.2, -0.1, 0.1, 0.3)
ssm_parameters(octant_cors)

}
\references{
Gurtman, M. B. (1992). Construct validity of interpersonal personality
measures: The interpersonal circumplex as a nomological net.
\emph{Journal of Personality and Social Psychology, 63}, 105-118.
}
