% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_power.R
\name{ssm_power_elevation}
\alias{ssm_power_elevation}
\title{Power for Single-Sample Elevation Test}
\usage{
ssm_power_elevation(effect, n, alpha = 0.05, k = 0.6, one_sided = FALSE)
}
\arguments{
\item{effect}{Expected absolute elevation}

\item{n}{Sample size}

\item{alpha}{Significance level (default: 0.05)}

\item{k}{Scaling constant for elevation SE (default: 0.60)}

\item{one_sided}{Use one-sided test? (default: FALSE)}
}
\value{
An object of class \code{ssm_power}
}
\description{
Calculates power for testing H0: elevation = 0 vs H1: elevation != 0
}
\examples{
ssm_power_elevation(effect = 0.11, n = 200)

}
