% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_power.R
\name{ssm_power_elevation_diff}
\alias{ssm_power_elevation_diff}
\title{Power for Two-Group Elevation Difference}
\usage{
ssm_power_elevation_diff(effect, n1, n2 = n1, alpha = 0.05, k = 0.6)
}
\arguments{
\item{effect}{Expected elevation difference}

\item{n1}{Sample size for group 1}

\item{n2}{Sample size for group 2 (default: same as n1)}

\item{alpha}{Significance level (default: 0.05)}

\item{k}{Scaling constant (default: 0.60)}
}
\value{
An object of class \code{ssm_power}
}
\description{
Calculates power for comparing elevations between two independent groups
}
\examples{
ssm_power_elevation_diff(effect = 0.11, n1 = 200, n2 = 200)

}
