% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_sample_size.R
\name{ssm_sample_size_elevation}
\alias{ssm_sample_size_elevation}
\title{Required Sample Size for Elevation Test}
\usage{
ssm_sample_size_elevation(
  effect,
  power = 0.8,
  alpha = 0.05,
  k = 0.6,
  one_sided = FALSE
)
}
\arguments{
\item{effect}{Expected absolute elevation}

\item{power}{Target power (default: 0.80)}

\item{alpha}{Significance level (default: 0.05)}

\item{k}{Scaling constant (default: 0.60)}

\item{one_sided}{Use one-sided test? (default: FALSE)}
}
\value{
An object of class \code{ssm_sample_size}
}
\description{
Calculates sample size needed to achieve target power for elevation test
}
\examples{
ssm_sample_size_elevation(effect = 0.11, power = 0.80)

}
